+++ database/insert/__init__.py
@@ -3,5 +3,6 @@ from database.insert.insert_1m_deposit_rates import *
 from database.insert.insert_asset import *
 from database.insert.insert_CPI import *
 from database.insert.insert_bond import *
-
+from database.insert.insert_fx import *
+from database.insert.insert_gdp import *
