+++ database/models/base.py
@@ -1,2 +1,3 @@
 from sqlalchemy.orm import declarative_base
+
 Base = declarative_base()
