+++ database/models/gdp/gdp_ref.py
@@ -1,10 +1,8 @@
-from sqlalchemy import Column, Integer, String, Float, Date, ForeignKey
-from sqlalchemy.orm import relationship, foreign
-from .base import Base
+from sqlalchemy import Column, Integer, ForeignKey, String
+from sqlalchemy.orm import relationship
 
-class GDP_Rates(Base):
-    __tablename__ = 'GDP_RATES'
-    country = Column(String(30), primary_key=True)
+from database.models.base import Base
+
 
 class GDP_Ref(Base):
     __tablename__ = "GDP_REF"
@@ -17,9 +15,3 @@ class GDP_Ref(Base):
         primaryjoin="GDP_Ref.series_id == foreign(Economic_Indicator.series_id)",
         uselist=False
     )
-
-class GDP_TS(Base):
-    __tablename__ = 'GDP_TS'
-    date = Column(Date, primary_key=True)
-    rate = Column(Float)
-    series_id = Column(Integer, ForeignKey('GDP_REF.series_id'), primary_key=True)
