+++ insert_data.py
@@ -1,4 +1,7 @@
 from database.insert import *
+from database.models import *
+
+
 
 # Provider
 #insert_new_provider('bloomberg2323222')
@@ -14,3 +17,13 @@ from database.insert import *
 
 # CPI
 #insert_all_cpi_currencies(provider_name='bloomberg')
+
+# FX
+#insert_all_fx_forward()
+#insert_all_spot_data()
+
+# GDP
+#insert_all_gdp()
+
+
+
