+++ old/insert_data_functions.py
@@ -1,6 +1,6 @@
 from database.session import session
 from database.models import *
-from functions import *
+from old.functions import *
 import pandas as pd
 from sqlalchemy import and_
 
@@ -149,7 +149,6 @@ def insert_fx(path,provider,duration):
 
 insert_yahoo_db("AAPL", start="2025-01-05", end="2025-01-15")
 insert_fx('AUDUSD.csv',1,'1M')
-
 list = ['USDCHF','USDSEK','USDNOK','USDAUD','USDNZD','USDGBP','USDEUR','USDJPY','USDCAD','CHFUSD','SEKUSD','NOKUSD','AUDUSD','NZDUSD','GBPUSD','EURUSD','JPYUSD','CADUSD']
